refactor: 패키지 구조를 그룹-단품-구성요소 순으로 정렬 변경
