+++ README.md
@@ -158,7 +158,8 @@ cards.add() 에서 불변처리 하자
 ## TODO
 
 - [ ] 이름들 변경
-- [ ] 테스트 코드 검사
+- [x] 패키지 구조 여러개-낱개-구성요소 순으로 정렬 변경
+- [x] 테스트 코드 검사
 - [ ] point를 객체로 두고 비교자 상속해 값 비교
 - [ ] 응집도를 높이는 방향으로 리팩토링
     - [ ] statistic map도 따로 클래스로 구현해보자, 아웃풋에 구현된 클래스를 넘겨줘 보자
