+++ src/test/java/blackjack/domain/card/CardDeckTest.java
@@ -3,7 +3,7 @@ package blackjack.domain.card;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.card.cards.CardDeck;
+import blackjack.domain.cards.CardDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
