+++ src/test/java/blackjack/domain/card/DenominationTest.java
@@ -2,7 +2,7 @@ package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.card.element.Denomination;
+import blackjack.domain.cards.card.denomination.Denomination;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
