+++ src/test/java/blackjack/domain/human/DealerTest.java
@@ -2,9 +2,10 @@ package blackjack.domain.human;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.element.Denomination;
-import blackjack.domain.card.element.Suit;
+import blackjack.domain.cards.card.Card;
+import blackjack.domain.cards.card.denomination.Denomination;
+import blackjack.domain.cards.card.denomination.Suit;
+import blackjack.domain.participant.human.Dealer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
