+++ src/test/java/blackjack/domain/human/NameTest.java
@@ -3,7 +3,7 @@ package blackjack.domain.human;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import blackjack.domain.human.element.Name;
+import blackjack.domain.participant.human.name.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
